+++ thrill/common/math.hpp
@@ -66,8 +66,9 @@ std::tuple<size_t, size_t> CalculateLocalRange(
 
     double per_pe = static_cast<double>(global_size) / static_cast<double>(p);
     return std::make_tuple(
-        std::ceil(i * per_pe),
-        std::min(static_cast<size_t>(std::ceil((i + 1) * per_pe)),
+        static_cast<size_t>(std::ceil(static_cast<double>(i) * per_pe)),
+        std::min(static_cast<size_t>(
+                     std::ceil(static_cast<double>(i + 1) * per_pe)),
                  global_size));
 }
